backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic-settings==2.2.1
apscheduler==3.10.4
google-ads==24.1.0
httpx==0.27.0
python-dotenv==1.0.1
alembic==1.13.1
